% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLRShiny2.R
\name{BLRShiny2}
\alias{BLRShiny2}
\title{Launch 'BLRShiny2' Interface}
\usage{
BLRShiny2()
}
\value{
Nothing
}
\description{
BLRShiny2() loads interactive user interface built using R 'shiny'.
}
\details{
Start BLRShiny2

The interactive user interface is to provide an easy way for binary logistic regression analysis and downloading relevant plot.
}
\examples{
if(interactive()){
library(rmarkdown)
BLRShiny2()
}
}
\keyword{BLRShiny2}
